pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "1.0.0"
description = ""
authors = ["Me"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = { extras = ["standard"], version = "^0.34.2" }
pydantic = "^2.11.4"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
starlette = "^0.46.2"
pyjwt = "^2.10.1"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
pre-commit = "^4.2.0"
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.7"
mypy = "^1.15.0"
pytest-cov = "^6.1.1"

[tool.poetry.scripts]
start-dev = "scripts.start_dev:main"
start = "scripts.start:main"
precommit = "scripts.precommit_runner:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint.main]
py-version = "3.11"
ignore = ["venv", ".venv", "tests", "scripts"]
max-line-length = 88

[tool.mypy]
python_version = "3.11"
files = "src"
exclude = "scripts|tests"
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
addopts = "-s --cov=src --cov-report=term --cov-report=html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
